requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Babel==4.0.0
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
pymysql==1.1.0
